+++ aiidalab_ispg/steps.py
@@ -183,6 +183,92 @@ class CodeSettings(ipw.VBox):
         )
 
 
+class SubmitWorkChainStepBase(ipw.VBox, WizardAppWidgetStep):
+    """Base class for workflow submission stesps"""
+
+    input_structure = traitlets.Union(
+        [traitlets.Instance(StructureData), traitlets.Instance(TrajectoryData)],
+        allow_none=True,
+    )
+    process = traitlets.Instance(WorkChainNode, allow_none=True)
+    disabled = traitlets.Bool()
+
+    def __init__(self, components=None, **kwargs):
+
+        self.submit_button = ipw.Button(
+            description="Submit",
+            tooltip="Submit the calculation with the selected parameters.",
+            icon="play",
+            button_style="success",
+            layout=ipw.Layout(width="auto", flex="1 1 auto"),
+            disabled=False,
+        )
+
+        self.submit_button.on_click(self._on_submit_button_clicked)
+
+        children = [self.submit_button]
+        if components is not None:
+            children = components + children
+
+        super().__init__(children=children)
+
+    def _on_submit_button_clicked(self, _):
+        self.submit_button.disabled = True
+        self.submit()
+
+    def submit(self):
+        """Must be overriden in the child class"""
+        pass
+
+    def _get_state(self):
+        # Process is already running.
+        if self.process is not None:
+            return self.State.SUCCESS
+        # Input structure not specified.
+        if self.input_structure is None:
+            return self.State.INIT
+        # Structure ready, but input parameters are invalid
+        if not self._validate_input_parameters():
+            return self.State.READY
+        return self.State.CONFIGURED
+
+    def _update_state(self, _=None):
+        self.state = self._get_state()
+
+    def _validate_input_parameters(self) -> bool:
+        """Should be overriden by the child class"""
+        return True
+
+    @traitlets.observe("input_structure")
+    def _observe_input_structure(self, change):
+        # self.set_trait("builder_parameters", self._default_builder_parameters())
+        self._update_state()
+        # self._set_num_mpi_tasks_to_default()
+
+    @traitlets.observe("state")
+    def _observe_state(self, change):
+        with self.hold_trait_notifications():
+            self.disabled = change["new"] not in (
+                self.State.READY,
+                self.State.CONFIGURED,
+            )
+            self.submit_button.disabled = change["new"] != self.State.CONFIGURED
+
+    @traitlets.observe("process")
+    def _observe_process(self, change):
+        self._update_state()
+
+    def can_reset(self):
+        "Do not allow reset while process is running."
+        return self.state is not self.State.ACTIVE
+
+    def reset(self):
+        with self.hold_trait_notifications():
+            self.process = None
+            self.input_structure = None
+
+
+# TODO: Subclass SubmitWorkChainStepBase
 class SubmitAtmospecAppWorkChainStep(ipw.VBox, WizardAppWidgetStep):
     """Step for submission of a bands workchain."""
